cleanup, move extractor to multilogs/mlutil
